Update village's action
