+++ src/scripts/village/actions/index.ts
@@ -1,143 +1,7 @@
 import * as ActionTypes from '../constants/ActionTypes'
-import {strToMessage} from '../util'
 import {village} from '../types'
 
-export interface SocketMessageReturnType<T> {
-  payload: T
-  type: ActionTypes.Socket.MESSAGE
-}
-
-export const socket = {
-  close: (event: CloseEvent): {event: CloseEvent, type: ActionTypes.Socket.CLOSE} => ({
-    event,
-    type: ActionTypes.Socket.CLOSE
-  }),
-  error: (event: Event): {event: Event, type: ActionTypes.Socket.ERROR} => ({
-    event,
-    type: ActionTypes.Socket.ERROR
-  }),
-  message: (payload: village.Payload) => {
-    if (typeof payload['@id'] === 'undefined') {
-      switch (payload.type) {
-        case village.PayloadType.nextGameInvitation: {
-          const action: SocketMessageReturnType<village.Payload$NextGameInvitation> = {
-            payload: {
-              ... payload as village.Payload$NextGameInvitation,
-              '@payload': village.PayloadType.nextGameInvitation
-            },
-            type: ActionTypes.Socket.MESSAGE
-          }
-
-          return action
-        }
-        case village.PayloadType.nextGameInvitationIsClosed: {
-          const action: SocketMessageReturnType<village.Payload$NextGameInvitationIsClosed> = {
-            payload: {
-              ... payload as village.Payload$NextGameInvitationIsClosed,
-              '@payload': village.PayloadType.nextGameInvitationIsClosed
-            },
-            type: ActionTypes.Socket.MESSAGE
-          }
-
-          return action
-        }
-        default:
-          throw Error('Unkonown Message')
-      }
-    }
-    const message = strToMessage(payload['@id'])
-
-    switch (message) {
-      case village.Message.boardMessage: {
-        const action: SocketMessageReturnType<village.Payload$BoardMessage> = {
-          payload: {
-            ... payload as village.Payload$BoardMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      case village.Message.chatMessage: {
-        const action: SocketMessageReturnType<village.Payload$ChatMessage> = {
-          payload: {
-            ... payload as village.Payload$ChatMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      case village.Message.errorMessage: {
-        const action: SocketMessageReturnType<village.Payload$ErrorMessage> = {
-          payload: {
-            ... payload as village.Payload$ErrorMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      case village.Message.flavorTextMessage: {
-        const action: SocketMessageReturnType<village.Payload$FlavorTextMessage> = {
-          payload: {
-            ... payload as village.Payload$FlavorTextMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      case village.Message.scrollMessage: {
-        const action: SocketMessageReturnType<village.Payload$ScrollMessage> = {
-          payload: {
-            ... payload as village.Payload$ScrollMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      case village.Message.systemMessage: {
-        const action: SocketMessageReturnType<village.Payload$SystemMessage> = {
-          payload: {
-            ... payload as village.Payload$SystemMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      case village.Message.voteMessage: {
-        const action: SocketMessageReturnType<village.Payload$VoteMessage> = {
-          payload: {
-            ... payload as village.Payload$VoteMessage,
-            '@payload': message
-          },
-          type: ActionTypes.Socket.MESSAGE
-        }
-
-        return action
-      }
-      default:
-        throw Error('Unkonown Message')
-    }
-  },
-  open: (event: Event): {event: Event, type: ActionTypes.Socket.OPEN} => ({
-    event,
-    type: ActionTypes.Socket.OPEN
-  }),
-  send: (payload: village.Payload): {payload: village.Payload, type: ActionTypes.Socket.SEND} => ({
-    payload,
-    type: ActionTypes.Socket.SEND
-  })
-}
+export * from './socket'
 
 export type ActivateNextButton = {
   type: ActionTypes.App.ACTIVATE_NEXT_BUTTON
@@ -313,10 +177,3 @@ export const tick = ({start, time}: {start: number, time: number}): Tick => ({
   time,
   type: ActionTypes.App.TICK
 })
-
-export type SocketClose = ReturnType<typeof socket.close>
-export type SocketError = ReturnType<typeof socket.error>
-export type SocketMessage = ReturnType<typeof socket.message>
-export type SocketOpen = ReturnType<typeof socket.open>
-export type SocketSend = ReturnType<typeof socket.send>
-
